ai for sotware engineering/week 3/requirements.txt
pandas
scikit-learn
numpy
tensorflow
matplotlib
spacy
streamlit
Pillow 